Changed from byte[] to short[] of which it should have been all along in order to accommodate wildcards of -1 (or 0xFFFF).

+++ CSMemEdit/Memory.cs
@@ -114,7 +114,7 @@ namespace CSMemEdit
             return bufferWritten;
         }
 
-        public IntPtr QuickSearch(uint lowerAddress, uint upperAddress, byte[] searchPattern)
+        public IntPtr QuickSearch(uint lowerAddress, uint upperAddress, short[] searchPattern)
         {
             uint addr = 0;
